charts {chartjs}	R Documentation

Draw charts

Description

These functions are used to draw the different types of graphs.

Usage

cjsBar(chartjs, labels, stacked = FALSE)

cjsLine(chartjs, labels)

cjsPie(chartjs, labels)

cjsDoughnut(chartjs, labels, cutout = 50)

cjsPolar(chartjs, labels)

cjsRadar(chartjs, labels)

Arguments

chartjs
a chartjs object

labels
a vector of characters

stacked
a logical, defaults to FALSE. If TRUE, bars will be stacked at each x value.

cutout
the optional percentage of the inner cutout for Doughnut charts (defaults to 50)

[Package chartjs version 0.3.3.9000 Index]