cjsSeries {chartjs}	R Documentation

Add a series to a chart

Description

Use this function to add a series to a chart.

Usage

cjsSeries(cjs, data, type, label, scale)

## S3 method for class 'cjs_bar'
cjsSeries(cjs, data, type = "bar", label = NULL,
  scale = NULL)

## S3 method for class 'cjs_line'
cjsSeries(cjs, data, type = "line", label = NULL,
  scale = NULL)

## S3 method for class 'cjs_horizontalBar'
cjsSeries(cjs, data, type = "horizontalBar",
  label = NULL, scale = NULL)

## S3 method for class 'cjs_radar'
cjsSeries(cjs, data, type = NULL, label = NULL,
  scale = NULL)

## S3 method for class 'cjs_pie'
cjsSeries(cjs, data, type = NULL, label = NULL,
  scale = NULL)

## S3 method for class 'cjs_polarArea'
cjsSeries(cjs, data, type = NULL, label = NULL,
  scale = NULL)

Arguments

cjs
a chartjs object

data
a vector of data

type
a character or NULL, the representation of the data,

label
a character, the name of the series

scale
an integer, the numeric position of the scale to attach the series to

color
an optional list of colours, see colortypes

palette
an optional RColorBrewer palette, defaults to Paired, see display.brewer.all

Methods (by class)

cjs_bar: Add series to a bar chart

cjs_line: Add series to a line chart

cjs_horizontalBar: Add series to a horizontal bar chart

cjs_radar: Add series to a radar chart

cjs_pie: Add series to a pie chart

cjs_polarArea: Add series to a polarArea chart

[Package chartjs version 0.3.3.9000 Index]