colortypes {chartjs}	R Documentation

Color types

Description

Allowed color types

Arguments

borderColor
for all charts

backgroundColor
for all charts

pointBorderColor
for Line, Radar and Scatter charts

pointBackgroundColor
for Line, Radar and Scatter charts

hoverBorderColor
for Bar, Pie, Doughnut and PolarArea charts

hoverBackgroundColor
for Bar, Pie, Doughnut and PolarArea charts

pointHoverBorderColor
for Line, Radar and Scatter charts

pointHoverBackgroundColor
for Line, Radar and Scatter charts

Details

When adding custom colours to a chart, the color argument should have the structure displayed in the example and ALL necessary colors must be given.

Examples

color <- list(background = "#FF0000",
              border = "#00FF00",
              hover = list(background = "#FF0000",
                           border = "#00FF00"),
              point = list(background = "#FF0000",
                           border = "#00FF00",
                           hover = list(background = "#FF0000",
                                        border = "#00FF00")))

[Package chartjs version 0.3.3.9000 Index]