options {chartjs}	R Documentation

Edit default options

Description

Allows editing of the chart options.

Usage

cjsOptions(chartjs, ...)

Arguments

chartjs
a chartjs object

...
named arguments, see Details

Details

The allowed values in ... can be found in the Chart.js documentation under Charts.default.global. You can also use this function to edit the animation and elements options by passing them as a list, or to edit the chart-specific options. To edit tooltips, legends or titles, use the corresponding functions.

[Package chartjs version 0.3.3.9000 Index]