shiny-chartjs {chartjs}	R Documentation

Shiny bindings for chartjs

Description

Shiny bindings for chartjs

Usage

chartjsOutput(outputId, width = "100%", height = "400px")

renderChartjs(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId
the widget id in the server part of the application

width
the width of the widget

height
the height of the widget

expr
an R expression that evaluates to a chartjs object

env
the environment in which to evaluate expr

quoted
a boolean, is expr a quoted expression ?

[Package chartjs version 0.3.3.9000 Index]